+++ src/components/map/filters/filters/MapFilter.vue
@@ -23,7 +23,7 @@
                 v-if="hasActiveFilter"
                 type="info"
             >
-                You have an active filter
+                You have applied filter
 
                 <common-button
                     link-color="error500"
@@ -103,6 +103,7 @@
                         #col2
                     >
                         <map-filter-box
+                            always-show-text
                             is-number
                             :model-value="store.filter.users?.lists ?? []"
                             strict
@@ -372,6 +373,19 @@
                 </common-button>
             </common-button-group>
         </template>
+        <map-filters-presets
+            v-else
+            endpoint-suffix="filters"
+            has-share
+            :max-presets="MAX_FILTERS"
+            :presets="store.filterPresets"
+            :refresh
+            :selected-preset="store.filter"
+            type="filter"
+            @create="createFilterPreset"
+            @reset="resetUserFilter"
+            @save="setUserFilter"
+        />
         <common-popup v-model="filterReset">
             <template #title>
                 Filters Reset
@@ -408,6 +422,7 @@ import CommonTabs from '~/components/common/basic/CommonTabs.vue';
 import CommonBlockTitle from '~/components/common/blocks/CommonBlockTitle.vue';
 import MapFilterBox from '~/components/map/filters/filters/MapFilterBox.vue';
 import { useStore } from '~/store';
+import { klona } from 'klona/json';
 import { backupUserFilter, setUserFilter } from '~/composables/fetchers/filters';
 import type { SelectItem } from '~/types/components/select';
 import type { RadarDataAirline } from '~/utils/backend/storage';
@@ -419,8 +434,12 @@ import CommonSelect from '~/components/common/basic/CommonSelect.vue';
 import type { VatsimEventData } from '~/server/api/data/vatsim/events';
 import CommonColor from '~/components/common/basic/CommonColor.vue';
 import CommonButton from '~/components/common/basic/CommonButton.vue';
-import { parseFilterAltitude } from '~/utils/shared';
+import { MAX_FILTERS, parseFilterAltitude } from '~/utils/shared';
 import CommonButtonGroup from '~/components/common/basic/CommonButtonGroup.vue';
+import MapFiltersPresets from '~/components/map/filters/MapFiltersPresets.vue';
+import type { UserFilter } from '~/utils/backend/handlers/filters';
+import { sendUserPreset } from '~/composables/fetchers';
+import equal from 'deep-equal';
 
 const store = useStore();
 const tab = ref('filter');
@@ -442,6 +461,12 @@ const dataStore = useDataStore();
 
 const { data: events } = await useLazyAsyncData('events', async () => {
     return $fetch<VatsimEventData>('/api/data/vatsim/events');
+});
+
+const { refresh } = await useLazyAsyncData('filters-presets', async () => {
+    await store.fetchFiltersPresets();
+
+    return true;
 });
 
 const getClosestEvents = computed(() => {
@@ -470,7 +495,7 @@ const getRouteEvents = computed<SelectItem[]>(() => {
 });
 
 const areFiltersEqual = computed(() => {
-    return JSON.stringify(store.filter) === JSON.stringify(store.activeFilter);
+    return equal(store.filter, store.activeFilter);
 });
 
 const hasActiveFilter = computed(() => {
@@ -506,7 +531,7 @@ const setEventRoutes = (id: number) => {
 };
 
 const applyFilter = () => {
-    setUserActiveFilter(structuredClone(toRaw(store.filter)));
+    setUserActiveFilter(klona(store.filter));
     store.getVATSIMData(true);
 };
 
@@ -582,6 +607,11 @@ const setRoutes = (routes: string[]) => {
     }
 
     previousRoutesLength = store.filter.airports!.routes!.length;
+};
+
+const createFilterPreset = async (name: string, json: UserFilter) => {
+    setUserFilter(await sendUserPreset(name, json, 'filters', () => createFilterPreset(name, json)));
+    await refresh();
 };
 </script>