+++ src/components/map/popups/MapPopupAtc.vue
@@ -155,6 +155,17 @@
                     </template>
                     View Stats
                 </common-button>
+                <common-button @click="copy.copy(`${ config.public.DOMAIN }/?atc=${ atc.callsign }`)">
+                    <template #icon>
+                        <share-icon/>
+                    </template>
+                    <template v-if="copy.copyState.value">
+                        Copied!
+                    </template>
+                    <template v-else>
+                        Link
+                    </template>
+                </common-button>
             </common-button-group>
         </template>
     </common-info-popup>
@@ -180,6 +191,7 @@ import { getVATSIMMemberStats } from '~/composables/data';
 import CommonBubble from '~/components/common/basic/CommonBubble.vue';
 import CommonSpoiler from '~/components/common/vatsim/CommonSpoiler.vue';
 import CommonFavoriteList from '~/components/common/vatsim/CommonFavoriteList.vue';
+import ShareIcon from '@/assets/icons/kit/share.svg?component';
 
 const props = defineProps({
     overlay: {
@@ -194,6 +206,8 @@ const mapStore = useMapStore();
 const dataStore = useDataStore();
 const map = inject<ShallowRef<Map | null>>('map')!;
 const tab = ref('info');
+const config = useRuntimeConfig();
+const copy = useCopyText();
 
 const atc = computed(() => {
     return findAtcByCallsign(props.overlay?.data.callsign);