+++ src/components/map/popups/MapPopupFlightInfo.vue
@@ -366,9 +366,9 @@ const { data: stats } = useLazyAsyncData(`stats-pilot-${ props.pilot.cid }`, () 
 
                     width: 100%;
                     height: 2px;
+                    border-radius: 4px;
 
                     background: $darkgray850;
-                    border-radius: 4px;
                 }
 
                 &::after {